Workarounding nitrate.py setting default log level to WARN
For some reason nitrate.py module changes global log level to WARN
when imported. It is used by tcms_plugin, so after using it, all the rest
of the plugins log next to nothing. It not only makes it hard to debug,
it also means that rpm_version_fetcher_plugin doesn't log versions,
which results in missing rpm versions in job runs in Jenkins.

In this patch we roll back to the old log level right after importing nitrate.

Change-Id: I7368def6cc60944fbadf63ca5b361eb79fabc867
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/10982
Tested-by: Katarzyna Jachim <kjachim@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Gadi Ickowicz <gickowic@redhat.com>
Reviewed-by: Ilia Meerovich <imeerovi@redhat.com>