rhevm_tools: simplify environment setup + pep8
Remove the prepareVmWithRhevm function, since it's used in this test
only and there is nicer way how to do that, which is used in other
tests. This change is needed for config file clenaup which is needed
for running the test with iscsi.

Also made low level storagedomains pep8 compliant. Reduces eye bleeding
by 1247%.

Change-Id: I3b37a4ec3bcb922820c80ed77357691e7f05c126
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/11590
Reviewed-by: Katarzyna Jachim <kjachim@redhat.com>
Tested-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>