RHEVM-1635 [java SDK] AttributeError: 'module' object has no attribute 'HostNICStatistics'
Juan explain: we use the name HostNicStatistics in Python SDK
and HostNICStatistics in the Java SDK. This is not by design,
but a side effect of the different algorithms that we use to build names
in the generators of both SDKs. Unfortunately we can't fix this
inconsistency, as it would break backwards compatibility.

Change-Id: I879ebcc01a81db287c8279bf55be2c838815a572
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/14200
Tested-by: Kobi Hakimi <khakimi@redhat.com>
Reviewed-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>