Fix ART CLI sending remove vm command twice
Ticket: rhevm-1727

This bug/issue occurs when using user level api i.e. Filter = True in
CLI.
Before actually performing the remove the CLI
performs a search to locate the VM, for example:

GET /ovirt-engine/api/vms?search=name%3Dblablabla
    Filter: True

But the RESTAPI server ignores the "search" parameter when the
"Filter: True" header is used, it will instead return all the
virtual machines that the user has permissions for. In this case it
will return the information corresponding to the only virtual
machine available in the system:

The the CLI sends the DELETE request to the resource returned
in the previous step.

In our infrastructure, we do cli command validation against the
auto-completion and that is dona via pexpect and sending the following
command "remove vm DEADBEAF \t \t "
This command actually runs and as a result it will remove the vm in case
there is only one vm in the system. Obviously the tester didn't expect
to run the command ....
This explains the first remove vm command sent.
The second one is expected one and it is done after the validation,
however this one will fail because the vm is already gone.

- For now, instead od using DEADBEAF string we send a fake uuid.
- some pep8/coding style fixes

Change-Id: Iae1fd326270c6b3d7cce70544e3e3b08f9cd8bb6
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/14566
Tested-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Ondra Machacek <omachace@redhat.com>
Reviewed-by: Kobi Hakimi <khakimi@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>