Golden environment inventory record
As part of the golden environment building task, we would like to
keep record regarding some objects to be used later on the golden
environment test runner task.
For now we keep the host list so the log collector can make use of this
info and collect the vdsm logs.

Change-Id: Idd2f1c6035aeefa9df73ca954e477b2c73efd32f
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/14886
Tested-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Nelly Credi <ncredi@redhat.com>