import images from glance like storage domains
- The imported image can be a treated as
  1- disk;
  2- template;
We have issue here, that we can't override image name, therefore we
use the disk_id to poll on the status of the disk imported.
Note that in 3.6 the name of disk can be overriden when the actual import
occurs and my life would be easier to keep the disk importaed under
track. However, introducing the glance image as a class would allow us
to keep better context regarding each instance of disk imported.

Change-Id: I1d9fb517c6e6c21beb599a69d3257d937f14f269
Reviewed-on: https://code.engineering.redhat.com/gerrit/42830
Tested-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Tareq Alayan <talayan@redhat.com>