Excluding modules not ready for GE with 'special requirements' tag
Modules included here are not GE ready, so they'll break the environment.
Before this patch these modules were excluded in the jenkins config files,
this patch will substitute that approach.

For storage tests 'not extra_reqs' should be included in the runner as
an expression (e.g: 'team="storage" and not extra_reqs').

Change-Id: Ifaa342e5b3c7723398b976fe984bf00374527964
Reviewed-on: https://code.engineering.redhat.com/gerrit/46486
Reviewed-by: Raz Tamir <ratamir@redhat.com>
Tested-by: Carlos Mestre Gonzalez <cmestreg@redhat.com>
Reviewed-by: Gilad Lazarovich <glazarov@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Reviewed-by: Nelly Credi <ncredi@redhat.com>