Core Api Java: support all_content parameter
pass all_content parameter through Java Api, in the getElmFromLink
function. Before the snapshots objects received from server where
missing initialization object.

This solution is far from complete, actually it assumes that if
all_content is given it will invoke list(max, all_content) and in find
it will invoke list(query, casesensitive, max, all_content)

A better solution would be:
To create a mechanism to select the best list function available and
according to the parameters given, however there are a lot of exceptions
and it will require a lot of changes and time.

More details in:
https://projects.engineering.redhat.com/browse/RHEVM-2157

Change-Id: I310299d4e2fbfb8639a532e73b49074952021464
Reviewed-on: https://code.engineering.redhat.com/gerrit/51450
Tested-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Nelly Credi <ncredi@redhat.com>
Reviewed-by: Tareq Alayan <talayan@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/51592