Art plugin: Skip test when bz network problem
If there is connection problem  with bz server in configuration of
plugin than it tries three times with 30 sec delay, after that it
disables plugin.

If there is connection problem when query bz server for bugs, it
tries it three times with 5 sec delay, after that skip test and
continue with another test case.

Change-Id: I0d6cd0d90daae9bd993340e86f88be079be4aecf
Reviewed-on: https://code.engineering.redhat.com/gerrit/51268
Tested-by: Petr Balogh <pbalogh@redhat.com>
Reviewed-by: Kobi Hakimi <khakimi@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Nelly Credi <ncredi@redhat.com>