stucked-main-thread: move traceback to debug log
it prints traceback in case the main thread is stucked for more than
240s. I saw cases where it was sleeping in some look and this handler
detected stucked thread, which is good, but it makes such job unstable
because we have post-build script which looks for tracebacks in console.
From this reason I move traceback to debug, and print warning to console
instead.

Change-Id: I412763774d93db97b26eb184b78ff7c7957a76fb
Reviewed-on: https://code.engineering.redhat.com/gerrit/61456
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Tested-by: Lukas Bednar <lbednar@redhat.com>