Infra adjustments to fit vm pools extension
These changes are a part of a bigger patch I originally pushed.
I divided the patch to several patches for the convinience of the
reviewer.
These changes include small ammendments to some low level functions,
the addition of a high_level/vmpools.py module (to fit logic of some
methods that were in low level for some reason) and a few small
adjustments to other tests that are influenced by these changes.

Change-Id: Icc4472f053318d704dfe480cc3444b4944391871
Reviewed-on: https://code.engineering.redhat.com/gerrit/67240
Tested-by: Sefi Litmanovich <slitmano@redhat.com>
Reviewed-by: Israel Pinto <ipinto@redhat.com>
Reviewed-by: Raz Tamir <ratamir@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>