Part 1 of new vm pools test implementation
These changes are a part of a bigger patch I originally pushed.
This patch contains the init, config and base class for the tests,
as well as the sanity test module and the fixtures folder with the aaa
properties files used to add qa.lab.tlv.redhat.com domain to the setup.

Change-Id: I4b0d668e1ca4706eaaae1a53a0c9ebb1ffeefc48
Reviewed-on: https://code.engineering.redhat.com/gerrit/67241
Tested-by: Sefi Litmanovich <slitmano@redhat.com>
Reviewed-by: Israel Pinto <ipinto@redhat.com>
Reviewed-by: Raz Tamir <ratamir@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>