Part 3 of new vm pools test implementation
These changes are a part of a bigger patch I originally pushed.
This patch contains 3 more test modules (one of the ideas of the new
vm pool test suite is to divide tests to different modules.
I also remove vm_pool_test.py which was the previous module that
contained all the previous tests.

Change-Id: I71272e993cb143608a52fc60a983b247cc0660bd
Reviewed-on: https://code.engineering.redhat.com/gerrit/67247
Tested-by: Sefi Litmanovich <slitmano@redhat.com>
Reviewed-by: Israel Pinto <ipinto@redhat.com>
Reviewed-by: Raz Tamir <ratamir@redhat.com>
Reviewed-by: Lukas Bednar <lbednar@redhat.com>