ssl_plugin: change method to get ca.crt
In version 4 is possible to get certificate from URL:
http://engine40.example.com/ovirt-engine/services/pki-resource?resource
=ca-certificate&format=X509-PEM-CA
But according to juan it's not reliable cause of admin of server may
change certificate of web server. So better idea is to get certificate
from TLS handshake. That will display the chain of certificates
presented by the web server. The last one between --BEGIN and --END is
the CA certificate.

Change-Id: If5ed3ecd22efe6f9dac1a2971e31287273fa7c30
Reviewed-on: https://code.engineering.redhat.com/gerrit/69342
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Petr Balogh <pbalogh@redhat.com>
Tested-by: Petr Balogh <pbalogh@redhat.com>