Avoid connection leak
Currently the HTTP proxy class uses a list of connections in order to
implement a pool. When request fails with an authentication error it
opens a new connection, and uses it. The previous (and failed)
connection is kept in the list and closed when the proxy is garbage
collected. In previous versions of the server this wasn't a problem,
because the server used to send a "Connection: close" after the
authentication error, so the failed connections where kept in the list,
but closed, and they didn't consume any resource other than memory. But
the new version of the server doesn't send this "Connection: close"
header. As a result the failed connections aren't closed, and they
consume file descriptors and sockets, in additio to the memory that they
already used to consume. As the list of connections isn't really used as
a pool, just to store dead connections, this patch solves the problem
removing this list, and making sure that the dead connections are closed
inmediately.

Change-Id: Ib009cd000aa34cc35dfc1379bbcebab6b5ba2c30
https://bugzilla.redhat.com/1324549
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/72685
Tested-by: Kobi Hakimi <khakimi@redhat.com>
Reviewed-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Kobi Hakimi <khakimi@redhat.com>