Network: Fix multi_host test
During a positive check, the for loop in helper.update_network_and_check_changes(...) function
uses the same NIC name (as it was retrieved by the first NIC index -> NIC name translation) for each query.
In cases where hosts have NICs with the same index but different names, this could lead to failure.

Change-Id: Ib73d35dc0a54d668f2b25ac9019c0bc45672ac6b
Reviewed-on: https://code.engineering.redhat.com/gerrit/100680
Tested-by: Mor Kalfon <mkalfon@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>