core: simplify confs
This change introduce a new defaults.yml which holds all constants, and
get rid of all specs and validator madness.

This change will allow us to remove all confs in the rhevm-jenkins repo
and instead to use the yaml file which represents the GE we are running
on.

This also change the way we run the tests, we need explicitly to specify
the logger.yaml we are working with and the xsd as well.

We have two dictionaries:
 1. ART_CONFIG - holds the defaults and some data from the GE yaml
 2. GE - a dict which represents the data of the GE

Change-Id: I2dbd55315a1c49637dcf4a615ba76ec748093b3f
Reviewed-on: https://code.engineering.redhat.com/gerrit/108094
Reviewed-by: Tareq Alayan <talayan@redhat.com>
Tested-by: Tareq Alayan <talayan@redhat.com>