Pytest 3.2 aignment and tier markers refactoring
https://github.com/pytest-dev/pytest/commit/10094a3f091ab64a608bc5107e91cafc9faa0c4c
Patch above which was merged  to pytest 3.1 and changed logic of Mark class.
This patch should fix issue with missing _arglist and also with polarion
marker.

In pytest 3.2 there will be new change with pristine non transfered markers
and it will be possible to distinguish between function and class marker.

But we would like to change this logic soon and move all markers on test cases
and use py.test option -m to select only tests marked with specifics markers.
We can use then expression like py.test -m virt and tier1 to run only virt
tests with tier1.

Next step will be that teams should decorate each testcase and do not use
markers on classes.

Change-Id: Iaa0100c9bb114b3ea5da32999602015393f3f206
Reviewed-on: https://code.engineering.redhat.com/gerrit/108466
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Tested-by: Petr Balogh <pbalogh@redhat.com>
Reviewed-by: Petr Balogh <pbalogh@redhat.com>