RHEVM-3127: infra: Add plugin to report leftovers between teams runs
This new plugin defines some hooks which will be called at each team
setup and teardown. From these hooks, we call methods from the
inventory module, which give us the ge-state in dict.

At the end of each package execution we store the state of the
leftovers of the previous package execution into a dictionary,
that will be writen by the leftovers plugin in a json file,
passed as parameter.

This file is supposed to be attached as an artifact to our jobs
in jenkins and also be sent as attachment to the email that is
sent from our production jobs.

Change-Id: I3aadd10e6e763bbda2e10d3c59f6b050e92f8a24
Reviewed-on: https://code.engineering.redhat.com/gerrit/113155
Reviewed-by: Lukas Bednar <lbednar@redhat.com>
Reviewed-by: Katerina Koukiou <kkoukiou@redhat.com>
Tested-by: Katerina Koukiou <kkoukiou@redhat.com>