Logger: fix storage if skipped
In case that test is skipped for example from bugzilla plugin, there is set
storage from previous parametrized run because parametrize for new storage
is also just skipped. In this case there was written bad storage type in
the logger. This patch access storage type item.callspec.params.

Also removed API in this logger because it's not used anymore.

Change-Id: Ibdd264adeeb27b5e8a6dddf55280442898a0ac81
Reviewed-on: https://code.engineering.redhat.com/gerrit/121921
Reviewed-by: Tareq Alayan <talayan@redhat.com>
Reviewed-by: Petr Balogh <pbalogh@redhat.com>
Tested-by: Petr Balogh <pbalogh@redhat.com>