Network: fix multiple_queue_nics to suspend and run the last VM
This patch porpose is to prevent failures of tests that use
RunVMOnce and run after mutliple_queue_nics in the tier.

The bug described in the following link:
https://bugzilla.redhat.com/show_bug.cgi?id=1507884

The proposed workaround (until BZ is resolved):
Change the test to use the last VM, which is currently not used by any other network tests.

Change-Id: I60a3edac6a4763ff12e4dec29b7f35d30521b36c
Reviewed-on: https://code.engineering.redhat.com/gerrit/121924
Reviewed-by: Ori Ben Sasson <obensass@redhat.com>
Tested-by: Mor Kalfon <mkalfon@redhat.com>
Reviewed-by: Israel Pinto <ipinto@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>