Storage matrix: skip tests which should not run
It is possible that you want to run just on one storage like NFS, and
if test were marked with @storages(("iscsi", "glusterfs")) matrix plugin
didn't parametrize this test with any storage. It led pytest to collect and
run this test without any storage without parametrized. So in this case we
need to skip this test with some reasonable message that this test is supposed
to run on differen storage. In this case we can track skipped test an will
know that this test running only on NFS and that we should also trigger
sometime on that storage.

TICKET:
https://projects.engineering.redhat.com/browse/RHEVM-3757

Change-Id: Ia3d8ea7f76a2c0023630e4ed28857e20309eacf5
Reviewed-on: https://code.engineering.redhat.com/gerrit/122135
Reviewed-by: Avihai Efrat <aefrat@redhat.com>
Tested-by: Petr Balogh <pbalogh@redhat.com>
Reviewed-by: Petr Balogh <pbalogh@redhat.com>